Refactor Printscript test files: remove unused imports and improve package structure
