+++ src/test/kotlin/com/ingsis/grupo10/printscript/PrintscriptApplicationConfigurationTests.kt
@@ -2,7 +2,6 @@ package com.ingsis.grupo10.printscript
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.test.context.TestPropertySource
 import kotlin.test.assertNotNull
 
 @SpringBootTest
