+++ src/test/kotlin/com/ingsis/grupo10/printscript/PrintscriptApplicationUnitTests.kt
@@ -1,7 +1,6 @@
 package com.ingsis.grupo10.printscript
 
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertNotNull
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.test.context.TestPropertySource
 import kotlin.test.assertTrue
